made resource model
