--- books/models.py
@@ -1,12 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import AbstractUser
-
-
-class User(AbstractUser):
-    location = models.CharField(blank=True, null=True, max_length=100)git 
-
-    def __repr__(self):
-        return f"<User username={self.username}>"
-
-    def __str__(self):
-        return self.username
