+++ django_freeshelf/settings.py
@@ -48,7 +48,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "django_extensions",
     "debug_toolbar",
-    "books",
+    "resources",
 ]
 
 MIDDLEWARE = [
