+++ resources/views.py
